Add documentation; make minor code adjustments
An explanation of each function has been added.

The number of tested values of p has been
increased from 50 to 101.

The code for computing the competitive ratio
has been moved into a separate function to
make ad hoc calculations easier.

The formatting of numbers in the output has
been modified.